fix cpp casting errors "Comparison of narrow type with wide type in loop condition" (#6533)

+++ src/c_api.cpp
@@ -1464,7 +1464,7 @@ int LGBM_DatasetCreateFromCSR(const void* indptr,
   }
   OMP_INIT_EX();
   #pragma omp parallel for num_threads(OMP_NUM_THREADS()) schedule(static)
-  for (int i = 0; i < nindptr - 1; ++i) {
+  for (int i = 0; i < static_cast<int>(nindptr - 1); ++i) {
     OMP_LOOP_EX_BEGIN();
     const int tid = omp_get_thread_num();
     auto one_row = get_row_fun(i);
@@ -1604,7 +1604,7 @@ int LGBM_DatasetCreateFromCSC(const void* col_ptr,
   }
   OMP_INIT_EX();
   #pragma omp parallel for num_threads(OMP_NUM_THREADS()) schedule(static)
-  for (int i = 0; i < ncol_ptr - 1; ++i) {
+  for (int i = 0; i < static_cast<int>(ncol_ptr - 1); ++i) {
     OMP_LOOP_EX_BEGIN();
     const int tid = omp_get_thread_num();
     int feature_idx = ret->InnerFeatureIndex(i);
